requirements.txt
# Core ML/DL Libraries
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.14.0

# Protobuf version constraint (TensorFlow/Transformers compatibility)
protobuf>=4.25.0,<5.0

# GenRL Framework (vendored locally in rgym_exp/vendor/genrl/)
# gensyn-genrl==0.1.9  # Commented out - using vendored v0.1.6 for debugging

# Reasoning Tasks (GitHub-only package)
git+https://github.com/open-thought/reasoning-gym.git

# Data Processing
pandas>=1.5.0
pyyaml>=6.0

# HuggingFace Integration
huggingface-hub>=0.16.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.65.0
